+++ framework/core/src/libkungfu/yijinjing/io/io.cpp
@@ -166,7 +166,7 @@ reader_ptr io_device::open_reader(const data::location_ptr &location, uint32_t d
   return r;
 }
 
-writer_ptr io_device::open_writer(uint32_t dest_id, uint32_t page_size) {
+writer_ptr io_device::open_writer(uint32_t dest_id, uint64_t page_size) {
   if (home_->mode != mode::REPLAY) {
     return std::make_shared<writer>(home_, dest_id, lazy_, publisher_, low_latency_, bus_, page_size);
   } else {
@@ -175,7 +175,7 @@ writer_ptr io_device::open_writer(uint32_t dest_id, uint32_t page_size) {
   }
 }
 
-writer_ptr io_device::open_writer_at(const data::location_ptr &location, uint32_t dest_id, uint32_t page_size) {
+writer_ptr io_device::open_writer_at(const data::location_ptr &location, uint32_t dest_id, uint64_t page_size) {
   if (home_->mode != mode::REPLAY) {
     return std::make_shared<writer>(location, dest_id, lazy_, publisher_, low_latency_, bus_, page_size);
   } else {
@@ -184,7 +184,7 @@ writer_ptr io_device::open_writer_at(const data::location_ptr &location, uint32_
   }
 }
 
-writer_ptr io_device::open_hookable_writer(uint32_t dest_id, const writer_hook_ptr &hook, uint32_t page_size) {
+writer_ptr io_device::open_hookable_writer(uint32_t dest_id, const writer_hook_ptr &hook, uint64_t page_size) {
   if (home_->mode != mode::REPLAY) {
     return std::make_shared<hookable_writer>(home_, dest_id, lazy_, publisher_, low_latency_, bus_, page_size, hook);
   } else {
@@ -194,7 +194,7 @@ writer_ptr io_device::open_hookable_writer(uint32_t dest_id, const writer_hook_p
 }
 
 writer_ptr io_device::open_hookable_writer_at(const data::location_ptr &location, uint32_t dest_id,
-                                              const writer_hook_ptr &hook, uint32_t page_size) {
+                                              const writer_hook_ptr &hook, uint64_t page_size) {
   if (home_->mode != mode::REPLAY) {
     return std::make_shared<hookable_writer>(location, dest_id, lazy_, publisher_, low_latency_, bus_, page_size, hook);
   } else {
